Check for PHP 5.3 before loading the rest of the plugin
