+++ wsuwp-content-syndicate.php
@@ -8,35 +8,16 @@ Author URI: https://web.wsu.edu/
 Version: 1.0.0
 */
 
-namespace WSU\ContentSyndicate;
-
 // If this file is called directly, abort.
 if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-add_action( 'plugins_loaded', 'WSU\ContentSyndicate\bootstrap' );
-/**
- * Loads the WSUWP Content Syndicate base.
- *
- * @since 1.0.0
- */
-function bootstrap() {
-	include_once __DIR__ . '/includes/class-wsu-syndicate-shortcode-base.php';
-
-	add_action( 'init', 'WSU\ContentSyndicate\activate_shortcodes' );
-}
-
-/**
- * Activates the shortcodes built in with WSUWP Content Syndicate.
- *
- * @since 1.0.0
- */
-function activate_shortcodes() {
-	include_once( dirname( __FILE__ ) . '/includes/class-wsu-syndicate-shortcode-json.php' );
-
-	// Add the [wsuwp_json] shortcode to pull standard post content.
-	new \WSU_Syndicate_Shortcode_JSON();
-
-	do_action( 'wsuwp_content_syndicate_shortcodes' );
+// This plugin uses namespaces and requires PHP 5.3 or greater.
+if ( version_compare( PHP_VERSION, '5.3', '<' ) ) {
+	add_action( 'admin_notices', create_function( '',
+		"echo '<div class=\"error\"><p>" . __( 'WSUWP Content Syndicate requires PHP 5.3 to function properly. Please upgrade PHP or deactivate WSUWP Content Syndicate.', 'wsuwp-content-syndicate' ) . "</p></div>';" ) );
+	return;
+} else {
+	include_once __DIR__ . '/includes/wsuwp-content-syndicate.php';
 }
